Extract tests into integration test
