+++ src/node/domains/core.rs
@@ -374,53 +374,3 @@ impl CoreDomainNode {
         self.rpc_handlers.subscribe_new_blocks().await.context("Failed to subscribe to new blocks")
     }
 }
-
-#[cfg(test)]
-mod tests {
-    #![allow(clippy::unwrap_used)]
-
-    use tempfile::TempDir;
-
-    use super::*;
-    use crate::farmer::CacheDescription;
-    use crate::node::{chain_spec, domains, Role};
-    use crate::{Farmer, Node, PlotDescription};
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn test_core() {
-        crate::utils::test_init();
-
-        let dir = TempDir::new().unwrap();
-        let core = ConfigBuilder::new().build();
-        let node = Node::dev()
-            .role(Role::Authority)
-            .system_domain(domains::ConfigBuilder::new().core(core))
-            .build(dir.path(), chain_spec::dev_config())
-            .await
-            .unwrap();
-        let (plot_dir, cache_dir) = (TempDir::new().unwrap(), TempDir::new().unwrap());
-        let farmer = Farmer::builder()
-            .build(
-                Default::default(),
-                &node,
-                &[PlotDescription::minimal(plot_dir.as_ref())],
-                CacheDescription::minimal(cache_dir.as_ref()),
-            )
-            .await
-            .unwrap();
-
-        node.system_domain()
-            .unwrap()
-            .core()
-            .unwrap()
-            .subscribe_new_blocks()
-            .await
-            .unwrap()
-            .next()
-            .await
-            .unwrap();
-
-        farmer.close().await.unwrap();
-        node.close().await.unwrap();
-    }
-}
